remove needless casts
Fixes:

warning: casting to the same type is unnecessary (`f64` -> `f64`)
   --> src/rrd_graph_new.rs:514:23
    |
514 |                     + (((t - start_time) as f64 * width) as f64) / time_span
    |                       ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ help: try: `((t - start_time) as f64 * width)`
    |
    = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#unnecessary_cast

warning: casting float literal to `f64` is unnecessary
   --> src/rrd_graph_new.rs:208:20
    |
208 |     while (range / (2.0 as f64).powi(l)) < 4.0 {
    |                    ^^^^^^^^^^^^ help: try: `2.0_f64`
    |
    = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#unnecessary_cast

+++ src/rrd_grid.rs
@@ -45,11 +45,11 @@ impl Component for ProxmoxRRDGrid {
                 let cw = 800;
                 let width = width.max(cw);
                 let padding = 6;
-                let mut cols = (width / cw) as usize;
+                let mut cols = width / cw;
                 if cols == 0 {
                     cols = 1;
                 }
-                let col_width = (width as usize - 2 * padding) / cols;
+                let col_width = (width - 2 * padding) / cols;
                 self.cols = cols;
                 self.col_width = col_width - padding;
                 true
